Refactor to get rid of module structure
